backend/requirements/base.txt
Django==4.2.15
pillow==11.0.0
requests==2.32.3
django-stubs==5.1.1
django-filter==24.3
django-environ==0.11.2
drf-spectacular==0.27.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
django-health-check==3.18.3
drf-standardized-errors==0.14.1
djangorestframework-simplejwt==5.3.1
